Add `exec` to the wrappers.
This is said to be a better way to write them: https://unix.stackexchange.com/q/279024/395855
While originally proposed as a way to propagate errors correctly in https://github.com/HolyBlackCat/quasi-msys2/issues/43#issuecomment-3321830946, it doesn't seem to actually affect propagation in our cases. Even though the default behavior is to not stop at the first error, the last return code is still returned from the shell script correctly.